Updates to docker-entrypoint.sh
This makes generation a bit better and produces artifacts closer to what
the GitHub action produces. Still don't fully understand how to control
the name of the database and have it be `org/repo` but this is good
enough for now. Realistically, we probably don't want to force that as
it is less portable.